Rewind ability implemented
Implemented the ability to rewind. However there technically is a leak in my memory management because I'm not freeing it from the list. This isn't the end of the world because it's my own memory management so it won't break things yet but I do need to figure out how to free memory...